Enable binary decoding.
There are a lot of fiddly details about transiting between different
data encodings. And the QR spec is a bit vague on some parts.
I've detailed the problems at the sister project
https://github.com/nu-book/zxing-cpp/issues/62

I've decided to patch this in a works-for-me kind of way:
unmarked encodings are treated as bianry and if that's not
good enough the higher layers have the source material
and can try decoding it themselves.